# apps/server/swagger/swagger.yaml
basePath: /api
definitions:
...
  internal_modules_auth.AuthResponse:
    properties:
      data:
        $ref: '#/definitions/internal_modules_auth.AuthResponseData'
      success:
        example: true
        type: boolean
    type: object
  internal_modules_auth.AuthResponseData:
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refreshToken:
        example: a1b2c3d4e5f6...
        type: string
      user:
        $ref: '#/definitions/internal_modules_auth.AuthUser'
    type: object
  internal_modules_auth.AuthUser:
    properties:
      email:
        example: user@example.com
        type: string
      emailVerified:
        example: false
        type: boolean
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: John Doe
        type: string
    type: object
  internal_modules_auth.ForgotPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
  internal_modules_auth.GenericResponse:
    properties:
      data:
        additionalProperties: {}
        type: object
      success:
        example: true
        type: boolean
    type: object
  internal_modules_auth.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: Password123
        maxLength: 50
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_modules_auth.ResetPasswordRequest:
    properties:
      newPassword:
        example: NewPassword123
        maxLength: 50
        minLength: 8
        type: string
      token:
        example: a1b2c3d4e5f6g7h8i9j0
        type: string
    required:
    - newPassword
    - token
    type: object
  internal_modules_auth.SignupRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: Password123
        maxLength: 50
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@coderz.space
    name: API Support
  description: Comprehensive bootcamp management platform API with multi-tenant architecture
    and role-based access control
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Coderz.space Bootcamp Management API
  version: "1.0"
paths:
...
  /v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset token (always returns success to prevent email
        enumeration)
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent (if email exists)
          schema:
            $ref: '#/definitions/internal_modules_auth.GenericResponse'
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties: true
            type: object
      summary: Request password reset
      tags:
      - Auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to receive authentication tokens
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/internal_modules_auth.AuthResponse'
        "401":
          description: Unauthorized - invalid credentials
          schema:
            additionalProperties: true
            type: object
      summary: Authenticate user
      tags:
      - Auth
  /v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using a valid reset token
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            $ref: '#/definitions/internal_modules_auth.GenericResponse'
        "400":
          description: Bad request - validation error or invalid/expired token
          schema:
            additionalProperties: true
            type: object
      summary: Reset password with token
      tags:
      - Auth
  /v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/internal_modules_auth.AuthResponse'
        "400":
          description: Bad request - validation error or email already exists
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...